requirements.txt
torch>=2.0
torchvision>=0.15
tqdm>=4.64
numpy>=1.23
Pillow>=9.0
